+++ Source/ImoutoRebirth.Navigator/ImoutoRebirth.Navigator/Utils/AsyncCommand.cs
@@ -9,6 +9,13 @@ public interface IAsyncCommand : ICommand
     bool CanExecute();
 }
 
+public interface IAsyncCommand<in TParamType> : ICommand
+{
+    Task ExecuteAsync(TParamType? parameter);
+
+    bool CanExecute(TParamType? parameter);
+}
+
 public class AsyncCommand : IAsyncCommand
 {
     public event EventHandler? CanExecuteChanged;
@@ -64,4 +71,47 @@ public class AsyncCommand : IAsyncCommand
         await ExecuteAsync();
     }
     #endregion
+}
+
+public class AsyncCommand<TParamType> : IAsyncCommand<TParamType> where TParamType : class
+{
+    public event EventHandler? CanExecuteChanged;
+
+    private bool _isExecuting;
+    private readonly Func<TParamType?, Task> _execute;
+    private readonly Func<TParamType?, bool>? _canExecute;
+
+    public AsyncCommand(
+        Func<TParamType?, Task> execute,
+        Func<TParamType?, bool>? canExecute = null)
+    {
+        _execute = execute;
+        _canExecute = canExecute;
+    }
+
+    public bool CanExecute(TParamType? parameter) => !_isExecuting && (_canExecute?.Invoke(parameter) ?? true);
+
+    public async Task ExecuteAsync(TParamType? parameter)
+    {
+        if (CanExecute(parameter))
+        {
+            try
+            {
+                _isExecuting = true;
+                await _execute(parameter);
+            }
+            finally
+            {
+                _isExecuting = false;
+            }
+        }
+
+        RaiseCanExecuteChanged();
+    }
+
+    private void RaiseCanExecuteChanged() => CanExecuteChanged?.Invoke(this, EventArgs.Empty);
+
+    bool ICommand.CanExecute(object? parameter) => CanExecute(parameter as TParamType);
+
+    async void ICommand.Execute(object? parameter) => await ExecuteAsync(parameter as TParamType);
 }